services/scraper.py
import httpx
from bs4 import BeautifulSoup
from urllib.parse import quote

INDIAN_KANOON_BASE = "https://indiankanoon.org"
HEADERS = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
}


def scrape_indian_kanoon(query: str, max_results: int = 5) -> list:
    search_url = f"{INDIAN_KANOON_BASE}/search/?formInput={quote(query)}"
    results = []

    try:
        with httpx.Client(headers=HEADERS, timeout=10) as client:
            response = client.get(search_url)
            response.raise_for_status()

        soup = BeautifulSoup(response.text, "html.parser")
        result_divs = soup.find_all("div", class_="result")

        for div in result_divs[:max_results]:
            title_tag = div.find("a")
            snippet_tag = div.find("p")

            if not title_tag:
                continue

            title = title_tag.get_text(strip=True)
            link = INDIAN_KANOON_BASE + title_tag.get("href", "")
            snippet = snippet_tag.get_text(strip=True) if snippet_tag else ""

            results.append({
                "title": title,
                "link": link,
                "snippet": snippet,
                "source": "Indian Kanoon"
            })

    except Exception as e:
        print(f"Indian Kanoon scraper error: {e}")

    return results