requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
groq==0.9.0
ollama==0.2.1
sentence-transformers==3.0.1
chromadb==0.5.3
striprtf==0.0.26
python-docx==1.1.2
pypdf==4.2.0
httpx==0.27.0
beautifulsoup4==4.12.3
datasets==2.20.0
tqdm==4.66.4